@@ -1,13 +1,14 @@
 from setuptools import setup
 from setuptools import find_packages
 
-version = "1.0"
+version = "2.0"
 
 install_requires = [
-    "certbot>=0.34.0",
-    "acme",
+    "certbot>=2.8.0",
+    "acme>=2.8.0",
     "setuptools",
     "requests",
+    "pytz",
 ]
 
 with open("README.md") as f:
@@ -22,7 +23,7 @@ setup(
     url="https://github.com/clazzor/certbot-dns-wedos",
     author="Clazzor",
     license="Apache License 2.0",
-    python_requires=">=3.7",
+    python_requires=">=3.8",
     classifiers=[
         "Development Status :: 5 - Production/Stable",
         "Environment :: Plugins",
@@ -31,12 +32,12 @@ setup(
         "Operating System :: POSIX :: Linux",
         "Programming Language :: Python",
         "Programming Language :: Python :: 3",
-        "Programming Language :: Python :: 3.7",
         "Programming Language :: Python :: 3.8",
         "Programming Language :: Python :: 3.9",
         "Programming Language :: Python :: 3.10",
         "Programming Language :: Python :: 3.11",
         "Programming Language :: Python :: 3.12",
+        "Programming Language :: Python :: 3.13",
         "Topic :: Internet :: WWW/HTTP",
         "Topic :: Security",
         "Topic :: System :: Installation/Setup",
