@@ -1,162 +1,164 @@
-from requests import Session, post
-from subprocess import run, PIPE
-from logging import getLogger
-from hashlib import sha1
-from shlex import split
-from json import dumps
-from time import strftime, localtime, sleep
-from certbot.plugins.dns_common import DNSAuthenticator, validate_file_permissions
-from certbot import errors
+import hashlib, pytz, re
+from datetime import datetime 
 
-logger = getLogger(__name__)
-URL = "https://api.wedos.com/wapi/json"
-TTL = 300
+import requests, json
+from typing import Any, Callable, Optional
+
+import logging 
+from certbot import errors
+from certbot.plugins.dns_common import CredentialsConfiguration, DNSAuthenticator
+
+logger = logging.getLogger(__name__)
+URL = 'https://api.wedos.com/wapi/json'
+
+
+def convertDomain(func: Callable[..., Any]) -> Callable[..., Any]:
+    def wrap(self, domain: str, validation_name: str, validation: str) -> Any:
+        regex = '([a-zA-Z0-9-]+)(\.[a-zA-Z]{2,5})?(\.[a-zA-Z]+$)'
+        pureDomain = re.search(regex,  domain).group(0)
+        subDomain = re.sub('\.'+regex, '', validation_name)
+        return func(self, pureDomain, subDomain, validation)
+    return wrap
+
 
 class _WedosClient():
-
     def __init__(self, username: str, password: str) -> None:
         self.url = URL
         self.username = username
-        self.password = password
-        self.ttl = TTL
-        self.session = Session()
+        self.password = hashlib.sha1(password.encode('ascii')).hexdigest()
+        self.session = requests.Session()
 
+    def _findTxtID(self, data: dict, validation: str) -> int:
+        if 'data' not in data['response']: return -1
+        if 'row' not in data['response']['data']: return -1
+        data = data['response']['data']['row']
 
-    def _findID(self, data: dict, key: str) -> int:
-        ID = -1
-        try:
-            for row in data['data']['row']:
-                if row['rdata'] == key:
-                    ID = row['ID']
-                    break
-            return ID
-        except:
-            return ID
+        for record in data:
+            if 'rdata' not in record: continue
+            if 'ID' not in record: continue
+            if record['rdata'] == validation: return record['ID']
+        return -1
 
-    def _clientConnect(self, command: str, params: dict = None, debug: str = "") -> post:
-        logger.debug(debug)
-        auth = (self.username + self.password + strftime('%H', localtime())).encode("ascii")
-
-        data = {
-            "user": self.username,
-            "auth": sha1(auth).hexdigest(),
-            "command": command
-            }
-
-        if type(params) is dict: data.update({"data": params})
-        data = dumps({"request": data})
-        response = self.session.post(self.url, data={'request': data})
-
+    def _handlerResponse(self, response: requests.post) -> dict:
         if not response.ok:
-            raise errors.PluginError(f"Cannot access Api Wedos, HTTP error code is {response.status_code}")
-
-        if response.json()["response"]["code"] >= 2000:
-            raise errors.PluginError(f"Error code received from Wedos, code is {response.json()['response']['code']}")
-
+            raise errors.PluginError('Cannot access the Wedos API, '
+                                    f'HTTP error code is {response.status_code}')
+        try:
+            response = response.json()
+        except json.JSONDecodeError as e:
+            raise errors.PluginError('Error occurred while parsing the response '
+                                     f'from Wedos API into JSON format: {e}')
+        if 'response' not in response:
+            raise errors.PluginError('Unknown error occurred while receiving '
+                                     'response from Wedos API.')
+        if 'code' not in response['response']:
+            raise errors.PluginError('Missing WAPI Error code in the '
+                                     'response from Wedos API.')
+        if response['response']['code'] >= 2000:
+            raise errors.PluginError('Error code received from Wedos API, The error '
+                                    f'code is {response["response"]["code"]}')
         return response
 
+    def _clientSend(self, command: str, requirement: dict = None) -> dict:
+        time = datetime.now(pytz.timezone('Europe/Prague')).strftime('%H')
+        auth = self.username + self.password + time
+        auth = hashlib.sha1(auth.encode('ascii')).hexdigest()
 
+        data = {
+            'user': self.username,
+            'auth': auth,
+            'command': command,
+            'data': requirement
+        }
+
+        data = {'request': json.dumps({'request': data})}
+        headers = {'Content-Type': 'application/x-www-form-urlencoded'}
+        response = self.session.post(self.url, data=data, headers=headers)
+
+        return self._handlerResponse(response)
+
+    @convertDomain
     def add_txt_record(self, domain: str, validation_name: str, validation: str) -> None:
-        if any(arg is None for arg in [domain, validation_name, validation]):
-            raise errors.Error("Missing important component, this is error from Certbot, not from plugin")
+        dns_row_add = {
+            'domain': domain,
+            'name': validation_name,
+            'type': 'TXT',
+            'ttl': 300,
+            'rdata': validation
+        }
 
-        record = {"domain": domain,
-                  "name": validation_name.replace('.' + domain, ''),
-                  "type": "TXT",
-                  "ttl": self.ttl,
-                  "rdata": validation}
+        self._clientSend('dns-row-add', dns_row_add)
+        self._clientSend('dns-domain-commit', {'name': domain})
 
-        self._clientConnect("dns-row-add", record, "Creating TXT record ..")
-        self._clientConnect("dns-domain-commit", {"name": domain}, "TXT Record was created!")
-        logger.debug("Changes was commit on Wedos")
-
+    @convertDomain
     def del_txt_record(self, domain: str, validation_name: str, validation: str) -> None:
-        if any(arg is None for arg in [domain, validation_name, validation]):
-            raise errors.Error("Missing important component, this is error from Certbot, not from plugin")
+        dns_records = self._clientSend('dns-rows-list', {'domain': domain})
+        txtID = self._findTxtID(dns_records,  validation)
 
-        response = self._clientConnect("dns-rows-list", {"domain": domain}, "Finding ID of TXT record ..")
-        id = self._findID(response.json()["response"],  validation)
-        if id == -1:
-            logger.warn("Couldn't find created TXT record, is recommended to check it.")
+        if txtID == -1:
+            logger.warn('Could not find the created TXT record. It is recommended to check it.')
             return
 
-        self._clientConnect("dns-row-delete", {"domain": domain, "row_id": id}, "Deleting created TXT record ..")
-        self._clientConnect("dns-domain-commit", {"name": domain}, "TXT record was deleted!")
-        logger.debug("Changes was commit on Wedos")
+        self._clientSend('dns-row-delete', {'domain': domain, 'row_id': txtID})
+        self._clientSend('dns-domain-commit', {'name': domain})
 
 
 class Authenticator(DNSAuthenticator):
-
-    description = "Obtain certificates for Wedos dns servers."
-
-    def __init__(self, *args: any, **kwargs: any) -> None:
-        super(Authenticator, self).__init__(*args, **kwargs)
-        self.credentials = None
-        self.arguments = {}
-
-        for arg in ["credentials", "user", "auth", "finalize"]:
-            if self.conf(arg): self.arguments[arg] = self.conf(arg)
-            else: self.arguments[arg] = None
-
+    description = 'Obtain certificates for Wedos dns servers.'
+    def __init__(self, *args: Any, **kwargs: Any) -> None:
+        super().__init__(*args, **kwargs)
+        self.credentials: Optional[CredentialsConfiguration] = None
 
     @classmethod
-    def add_parser_arguments(cls, add: callable) -> None:
-        super(Authenticator, cls).add_parser_arguments(add, default_propagation_seconds=30)
-        add("credentials", help="Wedos credentials INI file.")
-        add("user", help="User for Wedos API. (Overwrite INI file)")
-        add("auth", help="SHA1 Auth (Password) for Wedos API. (Overwrite INI file)")
-        add("finalize", help="Path or command that will run after everything is done. (Overwrite INI file)")
-
+    def add_parser_arguments(cls, add: Callable[..., None],
+                             default_propagation_seconds: int = 360) -> None:
+        super().add_parser_arguments(add, default_propagation_seconds)
+        add('credentials', help='Wedos credentials INI file.')
 
     def more_info(self) -> str:
-        return "This plugin uses certbot's dns-01 challenge to create and delete TXT records\
-                on a Wedos domain server, thanks to the API interface called WAPI provided by Wedos.\
-                With this plugin, you can make wildcard ssl."
+        return 'This plugin uses the Certbot DNS-01 challenge to create and delete' + \
+            ' TXT records on the Wedos domain server using the Wedos API called' + \
+            ' the WAPI. With this plugin, you can generate wildcard SSL certificates.'
 
-    def _get_credentials(self) -> None:
-        self._configure_file('credentials', "Wedos credentials INI file.")
-        validate_file_permissions(self.arguments["credentials"])
+    def _validate_credentials(self, credentials: CredentialsConfiguration) -> None:
+        user = credentials.conf('user')
+        auth = credentials.conf('auth')
+        propagation_seconds = self.conf('propagation-seconds')
 
-        self.credentials = self._configure_credentials(
-            "credentials",
-            "Wedos credentials INI file.",
-            {
-                "user": "User for Wedos API.",
-                "auth": "SHA1 Auth (Password) for Wedos API.",
-            },
-        )
+        if not user:
+            raise errors.PluginError('Missing parameter USER (email) for the Wedos API.'
+                                     ' [dns_wedos_user=john@example.com]')
+        if not auth:
+            raise errors.PluginError('Missing parameter AUTH (password) for the Wedos API.'
+                                     ' [dns_wedos_auth=ExamplePassword]')
+        if propagation_seconds < 300: 
+            raise errors.PluginError('Propagation seconds cannot be lower than 300 seconds.'
+                                     ' (Recommended propagation time is 420 seconds)')
+        if '@' not in user:
+            raise errors.PluginError('Wrong parameter USER (email) for the Wedos API.')
+        if len(auth) < 8:
+            raise errors.PluginError('Wrong parameter AUTH (password) for the Wedos API.')
 
     def _setup_credentials(self) -> None:
-        if self.arguments["credentials"]: self._get_credentials()
-
-        for arg in ["user", "auth", "finalize"]:
-            if not self.credentials or not self.credentials.conf(arg): continue
-            self.arguments[arg] = self.arguments[arg] or self.credentials.conf(arg)
-
-        if not self.arguments["user"]:
-            raise errors.PluginError("Missing or incorrect argument 'user' for WAPI.")
-        if not self.arguments["auth"]:
-            raise errors.PluginError("Missing or incorrect argument 'auth' for WAPI.")
-
+        self.credentials = self._configure_credentials(
+            'credentials',
+            'Wedos Credentials INI file.',
+            None,
+            self._validate_credentials
+        )
 
     def _perform(self, domain: str, validation_name: str, validation: str) -> None:
-        logger.debug("Starting ..")
-        domain = '.'.join(domain.split('.')[-2:])
-        self._get_wedos_client().add_txt_record(domain, validation_name, validation)
-        logger.debug("Waiting 5 minutes for applied changes on internet")
-        sleep(300)
+        self._get_wedos_client().add_txt_record(
+            domain, validation_name, validation
+        )
 
     def _cleanup(self, domain: str, validation_name: str, validation: str) -> None:
-        domain = '.'.join(domain.split('.')[-2:])
-        self._get_wedos_client().del_txt_record(domain, validation_name, validation)
-
-        logger.debug(f"Processing final command: {self.arguments['finalize']}")
-        if not self.arguments["finalize"]: return
-        process = run(split(self.arguments["finalize"]), shell=False, stderr=PIPE)
-
-        if process.returncode != 0:
-            logger.warn(f"Command: \"{self.arguments['finalize']}\" returned non-zero exit status\
-                        {process.returncode} with error message:\n{process.stderr.decode().strip()}")
+        self._get_wedos_client().del_txt_record(
+            domain, validation_name, validation
+        )
 
     def _get_wedos_client(self) -> _WedosClient:
-        return _WedosClient(self.arguments["user"], self.arguments["auth"])
+        return _WedosClient(
+            self.credentials.conf('user'),
+            self.credentials.conf('auth')
+        )
